chore: refactoring map-related code into a separate class
