+++ features/exploredash/src/main/res/drawable/ic_blue_circle.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="oval">
     <solid android:color="@color/dash_blue" />
-    <size android:width="24dp" android:height="24dp" />
+    <size android:width="14dp" android:height="14dp" />
 </shape>
