+++ features/exploredash/src/main/res/drawable/ic_merchant.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="41dp"
+    android:height="41dp"
+    android:viewportWidth="41"
+    android:viewportHeight="41">
+  <path
+      android:pathData="M20.5,18.5m-17.5,0a17.5,17.5 0,1 1,35 0a17.5,17.5 0,1 1,-35 0"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M22.789,11H16.2276L15.6898,14.0333L21.6058,14.044C24.5208,14.044 25.3813,15.0982 25.3597,16.8514C25.349,17.755 24.9618,19.2716 24.7897,19.7664C24.3379,21.0787 23.4236,22.5845 19.9601,22.5738H14.2055L13.6569,25.6178H20.1967C22.4986,25.6178 23.4881,25.3489 24.5208,24.8649C26.8226,23.8 28.1994,21.5304 28.748,18.5617C29.5762,14.1408 28.5544,11 22.789,11Z"
+      android:fillColor="#008DE4"/>
+  <path
+      android:pathData="M13.4094,16.7884C11.6884,16.7884 11.441,17.907 11.2797,18.5847C11.0645,19.4667 11,19.8324 11,19.8324H17.7119C19.4329,19.8324 19.6803,18.7138 19.8417,18.0361C20.0568,17.1541 20.1213,16.7884 20.1213,16.7884H13.4094Z"
+      android:fillColor="#008DE4"/>
+</vector>
