+++ features/exploredash/src/main/res/drawable/user_location_map_marker.xml
@@ -1,16 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
+    <item android:top="1dp" android:left="1dp">
         <shape android:shape="oval">
-            <solid android:color="@color/white" />
-            <size android:width="24dp" android:height="24dp" />
+            <solid android:color="@color/shadow_color" />
+            <size android:width="21dp" android:height="21dp" />
         </shape>
     </item>
 
-    <item android:top="3dp"
-        android:left="3dp"
-        android:right="3dp"
-        android:bottom="3dp"
-        android:drawable="@drawable/ic_blue_circle"/>
+    <item>
+        <shape android:shape="oval">
+            <stroke android:width="1dp" android:color="@android:color/transparent" />
+            <solid android:color="@color/white" />
+            <size android:width="21dp" android:height="21dp" />
+        </shape>
+    </item>
 
+    <item android:top="3.5dp"
+        android:left="3.5dp"
+        android:right="3.5dp"
+        android:bottom="3.5dp"
+        android:drawable="@drawable/ic_blue_circle"/>
 </layer-list>
